requirements.txt
pyzmq==15.2.0
locustio==0.7.3
thriftpy==0.3.8

